requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
deepspeed==0.12.6
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
latex2mathml==3.77.0
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
peft==0.10.0
pillow==10.4.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynvml==11.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.7
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.27.0
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.37.2
triton==2.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.18.0
wavedrom==2.0.3.post3
websockets==11.0.3
flash-attn==2.5.7
